✨ Refactor mind provider architecture
